requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
readability-lxml>=0.8.1
pyyaml>=6.0.1
rich>=13.7.0
lxml>=5.1.0
